services/historical_replayer/requirements.txt
confluent-kafka
psycopg2-binary
python-dotenv
numpy
argparse
